+++ internal/eventhub/variables_test.go
@@ -1,9 +1,8 @@
-package handlers
+package eventhub
 
 import (
 	"context"
 	"testing"
-	"time"
 
 	"github.com/decisiveai/mdai-data-core/eventing"
 	"github.com/stretchr/testify/assert"
@@ -14,8 +13,6 @@ import (
 type MockHandlerAdapter struct {
 	Calls map[string][]map[string]string
 }
-
-var _ IHandlerAdapter = (*MockHandlerAdapter)(nil)
 
 func (mh *MockHandlerAdapter) AddElementToSet(_ context.Context, variableKey, hubName, value, correlationID string) error {
 	mh.Calls["AddElementToSet"] = append(mh.Calls["AddElementToSet"], map[string]string{
@@ -37,7 +34,7 @@ func (mh *MockHandlerAdapter) RemoveElementFromSet(_ context.Context, variableKe
 	return nil
 }
 
-func (mh *MockHandlerAdapter) AddSetMapElement(_ context.Context, variableKey, hubName, field, value, correlationID string) error {
+func (mh *MockHandlerAdapter) SetMapEntry(_ context.Context, variableKey, hubName, field, value, correlationID string) error {
 	mh.Calls["AddSetMapElement"] = append(mh.Calls["AddSetMapElement"], map[string]string{
 		"variableKey":   variableKey,
 		"hubName":       hubName,
@@ -48,7 +45,7 @@ func (mh *MockHandlerAdapter) AddSetMapElement(_ context.Context, variableKey, h
 	return nil
 }
 
-func (mh *MockHandlerAdapter) RemoveElementFromMap(_ context.Context, variableKey, hubName, field, correlationID string) error {
+func (mh *MockHandlerAdapter) RemoveMapEntry(_ context.Context, variableKey, hubName, field, correlationID string) error {
 	mh.Calls["RemoveElementFromMap"] = append(mh.Calls["RemoveElementFromMap"], map[string]string{
 		"variableKey":   variableKey,
 		"hubName":       hubName,
@@ -67,19 +64,6 @@ func (mh *MockHandlerAdapter) SetStringValue(_ context.Context, variableKey, hub
 	})
 	return nil
 }
-
-// stubDeps implements VarDeps for tests.
-type stubDeps struct {
-	logger  *zap.Logger
-	adapter IHandlerAdapter
-}
-
-var _ VarDeps = (*stubDeps)(nil)
-
-func (s *stubDeps) GetLogger() *zap.Logger { return s.logger }
-
-//nolint:ireturn // VarDeps intentionally abstracts via interface for tests and decoupling
-func (s *stubDeps) GetHandlerAdapter() IHandlerAdapter { return s.adapter }
 
 func TestHandleManualVariablesActions(t *testing.T) {
 	t.Parallel()
@@ -229,180 +213,10 @@ func TestHandleManualVariablesActions(t *testing.T) {
 		t.Run(tc.description, func(t *testing.T) {
 			t.Parallel()
 			mockHandlerAdapter := &MockHandlerAdapter{Calls: make(map[string][]map[string]string)}
-			mdai := &stubDeps{logger: zap.NewNop(), adapter: mockHandlerAdapter}
+			mdai := &VarDeps{Logger: zap.NewNop(), HandlerAdapter: mockHandlerAdapter}
 
-			require.NoError(t, HandleManualVariablesActions(t.Context(), mdai, tc.event))
+			require.NoError(t, mdai.HandleManualVariablesActions(t.Context(), tc.event))
 			assert.Contains(t, mockHandlerAdapter.Calls[tc.handlerName], tc.expected)
 		})
 	}
-}
-
-func TestBuildSlackPayload(t *testing.T) {
-	t.Parallel()
-
-	testCases := []struct {
-		desc     string
-		args     map[string]string
-		event    eventing.MdaiEvent
-		payload  map[string]any
-		expected SlackPayload
-	}{
-		{
-			desc: "build minimal slack payload",
-			args: map[string]string{},
-			event: eventing.MdaiEvent{
-				HubName:   "foobar",
-				Name:      "barbaz",
-				Timestamp: time.Date(2021, time.September, 21, 9, 21, 9, 21, time.UTC),
-			},
-			payload: map[string]any{
-				"status": "whoa",
-				"labels": map[string]any{},
-			},
-			expected: SlackPayload{
-				Text: "MDAI Hub Event - foobar - barbaz",
-				Blocks: []map[string]any{
-					{
-						"type": "section",
-						"text": map[string]string{
-							"type": "mrkdwn",
-							"text": "*MDAI Hub Event - foobar - barbaz*",
-						},
-					},
-					{
-						"type": "section",
-						"fields": []map[string]string{
-							{
-								"type": "mrkdwn",
-								"text": "*Alert timestamp* - 2021-09-21 09:21:09.000000021 +0000 UTC",
-							},
-							{
-								"type": "mrkdwn",
-								"text": "*alertname* - Unknown",
-							},
-							{
-								"type": "mrkdwn",
-								"text": "*status* - whoa",
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			desc: "build more complex slack payload",
-			args: map[string]string{
-				"labels_val_ref_primary": "lol",
-				"message":                "SLACKY MCSLACKFACE LOL",
-				"link_text":              "CLICK HERE FOR FREE IPAD!",
-				"link_url":               "https://www.example.com",
-			},
-			event: eventing.MdaiEvent{
-				HubName:   "foobaz",
-				Name:      "barbar",
-				Timestamp: time.Date(2021, time.September, 21, 9, 21, 9, 21, time.UTC),
-			},
-			payload: map[string]any{
-				"status": "whoa",
-				"labels": map[string]any{
-					"lol":       "wut",
-					"alertname": "k.",
-				},
-			},
-			expected: SlackPayload{
-				Text: "SLACKY MCSLACKFACE LOL",
-				Blocks: []map[string]any{
-					{
-						"type": "section",
-						"text": map[string]string{
-							"type": "mrkdwn",
-							"text": "*SLACKY MCSLACKFACE LOL*",
-						},
-					},
-					{
-						"type": "section",
-						"fields": []map[string]string{
-							{
-								"type": "mrkdwn",
-								"text": "*Alert timestamp* - 2021-09-21 09:21:09.000000021 +0000 UTC",
-							},
-							{
-								"type": "mrkdwn",
-								"text": "*lol* - wut",
-							},
-							{
-								"type": "mrkdwn",
-								"text": "*alertname* - k.",
-							},
-							{
-								"type": "mrkdwn",
-								"text": "*status* - whoa",
-							},
-						},
-					},
-					{
-						"type": "actions",
-						"elements": []map[string]any{
-							{
-								"type": "button",
-								"text": map[string]string{
-									"type": "plain_text",
-									"text": "CLICK HERE FOR FREE IPAD!",
-								},
-								"style": "primary",
-								"url":   "https://www.example.com",
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.desc, func(t *testing.T) {
-			t.Parallel()
-			actual, err := buildSlackPayload(tc.args, tc.event, tc.payload)
-			require.NoError(t, err)
-			assert.Equal(t, tc.expected, actual)
-		})
-	}
-}
-
-func TestProcessEventPayload_Success(t *testing.T) {
-	validJSON := `{"key1":"value1","key2":42,"nested":{"sub":"val"}}`
-	event := eventing.MdaiEvent{
-		ID:      "e1",
-		Name:    "test",
-		HubName: "hub1",
-		Payload: validJSON,
-	}
-
-	result, err := ProcessEventPayload(event)
-	require.NoError(t, err, "expected no error for valid JSON payload")
-
-	assert.Contains(t, result, "key1")
-	assert.Contains(t, result, "key2")
-	assert.Contains(t, result, "nested")
-
-	assert.Equal(t, "value1", result["key1"])
-	assert.InDelta(t, float64(42), result["key2"], 0.001)
-	nested, ok := result["nested"].(map[string]any)
-	assert.True(t, ok, "expected nested to be a map[string]any")
-	assert.Equal(t, "val", nested["sub"])
-}
-
-func TestProcessEventPayload_InvalidJSON(t *testing.T) {
-	invalidJSON := `{"key1":"value1", "key2":}`
-	event := eventing.MdaiEvent{
-		ID:      "e2",
-		Name:    "test-invalid",
-		HubName: "hub2",
-		Payload: invalidJSON,
-	}
-
-	result, err := ProcessEventPayload(event)
-	assert.Nil(t, result, "expected result to be nil on invalid JSON")
-	require.Error(t, err, "expected an error for invalid JSON payload")
-	assert.Contains(t, err.Error(), "failed to unmarshal payload")
 }
